fix(omsapi): decode avg/lead-time fields as float64 (backend sends float averages)
Starting a new PO from a supplier with reorder-queue items crashed with
"cannot unmarshal number 3.0 into ... avg_lead_time of type int". The backend
computes lead-time as an average (a JSON float), but the Go structs typed these
as int. Changed all three (ReorderDataSupplier.AvgLeadTime, ItemSupplier
average_lead_time, the reorder-item AverageLeadTime) int -> float64; updated 4
display sites %dd -> %gd; added regression tests. Same class the sc-8gy sweep
missed (int-typed computed average).

+++ internal/omsapi/inventory.go
@@ -72,7 +72,7 @@ type Item struct {
 	UnitCost             DecimalString `json:"unit_cost,omitempty"`
 	PackageCost          DecimalString `json:"package_cost,omitempty"`
 	QuantityPerPackage   int           `json:"quantity_per_package,omitempty"`
-	AverageLeadTime      int           `json:"average_lead_time,omitempty"`
+	AverageLeadTime      float64       `json:"average_lead_time,omitempty"`
 	TotalValue           DecimalString `json:"total_value,omitempty"`
 	ThumbnailURL         string        `json:"thumbnail,omitempty"`
 	QRCodeURL            string        `json:"qr_code_url,omitempty"`
@@ -834,7 +834,7 @@ type ItemSupplier struct {
 	PackQuantity      int           `json:"quantity_per_package,omitempty"`
 	UnitCost          DecimalString `json:"unit_cost,omitempty"`
 	PackageCost       DecimalString `json:"package_cost,omitempty"`
-	LeadTimeDays      int           `json:"average_lead_time,omitempty"`
+	LeadTimeDays      float64       `json:"average_lead_time,omitempty"`
 	IsPreferred       bool          `json:"is_primary,omitempty"`
 	IsActive          bool          `json:"is_active,omitempty"`
 	IsDiscontinued    bool          `json:"is_discontinued,omitempty"`
